import similar question tests
remove unused blanks
