+++ openprocurement/tender/openeu/tests/question.py
@@ -3,21 +3,30 @@ import unittest
 
 from openprocurement.api.tests.base import snitch
 
+from openprocurement.tender.belowthreshold.tests.question_blanks import (
+    # TenderQuestionResourceTest
+    create_tender_question_invalid,
+    get_tender_question,
+    get_tender_questions,
+    # TenderLotQuestionResourceTest
+    lot_create_tender_question,
+    lot_patch_tender_question,
+)
+
 from openprocurement.tender.openeu.tests.base import (
     BaseTenderContentWebTest,
     test_bids,
     test_lots,
 )
-from openprocurement.tender.openeu.tests.questions_blanks import (
-    # TenderLotQuestionResourceTest
-    lot_create_tender_question,
-    lot_patch_tender_question,
+from openprocurement.tender.openeu.tests.question_blanks import (
     # TenderQuestionResourceTest
-    create_tender_question_invalid,
-    create_tender_question,
     patch_tender_question,
-    get_tender_question,
-    get_tender_questions,
+    answering_question,
+)
+
+from openprocurement.tender.openua.tests.question_blanks import (
+    # TenderQuestionResourceTest
+    create_tender_question,
 )
 
 
@@ -25,10 +34,12 @@ class TenderQuestionResourceTest(BaseTenderContentWebTest):
 
     initial_auth = ('Basic', ('broker', ''))
     test_bids_data = test_bids
+    test_tender_status = 'active.pre-qualification'
 
     test_create_tender_question_invalid = snitch(create_tender_question_invalid)
     test_create_tender_question = snitch(create_tender_question)
     test_patch_tender_question = snitch(patch_tender_question)
+    test_answering_question = snitch(answering_question)
     test_get_tender_question = snitch(get_tender_question)
     test_get_tender_questions = snitch(get_tender_questions)
